# process walker perfs: optimize readLimits and readStats