… probe/process/cache.go