- optimize readStats
- optimize readLimits
- cache limits and cmdline/name after parsing

Fixes https://github.com/weaveworks/scope/issues/2461