@2opremio 2017-05-03
@alban I don't doubt that the code is much faster now but it's very hard to read. Please try to refactor it. I think that separating the parsing of integers from the space-counting code would already be an improvement.

Also, are you sure that parsing ints manually provides an improvement? I wouldn't be surprised if the Golang library wasn't optimized in this respect.