@alban 2017-05-03
The issue with [strconv.Atoi](https://golang.org/pkg/strconv/#Atoi) or [strconv.ParseInt](https://golang.org/pkg/strconv/#ParseInt) is that they don't stop parsing when finding a non-digit such as a space and they return an error.

The previous code was using [strings.Fields](https://golang.org/pkg/strings/#Fields) to allocate and build a string compatible with [strconv.Atoi](https://golang.org/pkg/strconv/#Atoi) (containing only the digits without extra spaces). The pprof graphs showed that it was [strings.Fields](https://golang.org/pkg/strings/#Fields) taking the cpu rather than [strconv.Atoi](https://golang.org/pkg/strconv/#Atoi) but we cannot use Atoi without Fields.

`fmt.Sscanf("%d "...)` is able to work with extra spaces but it's also slower.

I'll try to refactor this.