@alban 2017-05-03
Code refactored and rebased.

Now that #2456 is merged in master, I ran performance tests again to see if this PR was still bringing better perfs. I used the test scenario described in https://github.com/weaveworks/scope/issues/2461#issuecomment-297036674

## Master branch

Up to 95660e3da735634d566325a43ab33c4380b8d346 (so #2456 is included)

![pprof-branch-master](https://cloud.githubusercontent.com/assets/27575/25669394/aace8faa-3029-11e7-9986-bf6e1c45535a.png)

## This PR

![pprof-branch-2491](https://cloud.githubusercontent.com/assets/27575/25669401/b1f95c60-3029-11e7-9114-4b1d8c58fd2e.png)

## Conclusion

The function `github.com/weaveworks/scope/probe/process.(*walker).Walk` takes respectively:
- 15.32s
- 12.91s (16% faster) 

Overall, the graphs' titles say respectively:
- 36.37s
- 31.32s (19% faster overall, probably coincidental)